+++ asyncdex/models/custom_list.py
@@ -2,7 +2,6 @@ from typing import Any, Dict
 
 from .abc import Model
 from .manga_list import MangaList
-from .mixins import DatetimeMixin
 from .user import User
 from ..constants import routes
 from ..enum import Visibility