+++ docs/changelog.rst
@@ -49,10 +49,13 @@ Added
 * :meth:`.Manga.delete`
 * :meth:`.add_to_list`
 * :meth:`.remove_from_list`
-* :meth:`.follow`
-* :meth:`.unfollow`
+* :meth:`.Manga.follow`
+* :meth:`.Manga.unfollow`
 * :meth:`.TitleList.parts`
 * :meth:`.create_manga`
+* :meth:`.create_author`
+* :meth:`.Author.update`
+* :meth:`.Author.delete`
 
 Changed
 +++++++